Remove `message` template
