--- apps/meteor/app/oembed/client/baseWidget.html
@@ -1,3 +1,0 @@
-<template name="oembedBaseWidget">
-	{{> Template.dynamic template=template }}
-</template>
