--- apps/meteor/app/oembed/client/index.ts
@@ -1,11 +1,0 @@
-import './baseWidget.html';
-import './baseWidget';
-import './oembedImageWidget.html';
-import './oembedImageWidget';
-import './oembedAudioWidget.html';
-import './oembedVideoWidget.html';
-import './oembedVideoWidget';
-import './oembedYoutubeWidget.html';
-import './oembedUrlWidget.html';
-import './oembedUrlWidget';
-import './oembedFrameWidget.html';
