--- apps/meteor/app/oembed/client/oembedAudioWidget.html
@@ -1,16 +1,0 @@
-<template name="oembedAudioWidget">
-	<a href="{{url}}" rel="noopener noreferrer" target="_blank">
-		{{#if parsedUrl}}
-			{{> collapseArrow collapsedMedia=collapsedMediaVar}}
-			{{#unless collapsed}}
-				<span class="collapse-switch icon-down-dir" data-index="{{index}}" data-collapsed="{{collapsed}}"></span>
-				<blockquote class="message-oembed background-transparent-darker-before">
-					<audio controls>
-						<source src="{{url}}" rel="noopener noreferrer" type="{{headers.contentType}}">
-						Your browser does not support the audio element.
-					</audio>
-				</blockquote>
-			{{/unless}}
-		{{/if}}
-	</a>
-</template>
