--- apps/meteor/app/oembed/client/oembedFrameWidget.html
@@ -1,30 +1,0 @@
-<template name="oembedFrameWidget">
-{{#if parsedUrl}}
-	<blockquote class="message-oembed background-transparent-darker-before">
-		{{#if meta.oembedProviderName}}
-			{{#if meta.oembedProviderUrl}}
-				<a href="{{meta.oembedProviderUrl}}" rel="noopener noreferrer" target="_blank" style="color: #9e9ea6">{{meta.oembedProviderName}}</a>
-			{{/if}}
-		{{/if}}
-		{{#if meta.oembedAuthorName}}
-			{{#if meta.oembedAuthorUrl}}
-				<br class="only-after-a"/>
-				<a href="{{meta.oembedAuthorUrl}}" rel="noopener noreferrer" target="_blank">{{meta.oembedAuthorName}}</a>
-			{{/if}}
-		{{/if}}
-		{{#if meta.oembedTitle}}
-			{{#if meta.oembedUrl}}
-				<br class="only-after-a"/>
-				<a href="{{meta.oembedUrl}}" rel="noopener noreferrer" target="_blank">{{meta.oembedTitle}}</a>
-			{{/if}}
-		{{/if}}
-		{{> collapseArrow collapsedMedia=collapsedMediaVar}} <br>
-		{{#unless collapsed}}
-			{{#if meta.oembedDescription}}
-				<p>{{meta.oembedDescription}}</p>
-			{{/if}}
-			{{{meta.oembedHtml}}}
-		{{/unless}}
-	</blockquote>
-{{/if}}
-</template>
