--- apps/meteor/app/oembed/client/oembedUrlWidget.html
@@ -1,32 +1,0 @@
-<template name="oembedUrlWidget">
-	{{#if show}}
-		<blockquote class="message-oembed background-transparent-darker-before">
-			<div style="{{#if image}}min-height: 80px;{{/if}} padding: 10px 3px;">
-				{{#if image}}
-					{{#if meta.ogImageUserGenerated}}
-						<div>
-							<a href="{{url}}" rel="noopener noreferrer" target="{{target}}">
-								{{#if meta.ogImage}}
-									<img src="{{meta.ogImage}}" rel="noopener noreferrer" height="200" />
-								{{else}}
-									<img src="{{image}}" rel="noopener noreferrer" height="200" />
-								{{/if}}
-							</a>
-						</div>
-					{{else}}
-						<a href="{{url}}" rel="noopener noreferrer" target="{{target}}">
-							<img src="{{image}}" rel="noopener noreferrer" height="60" align="left" style="margin-right: 10px;"/>
-						</a>
-					{{/if}}
-				{{/if}}
-				<div style="color: #AAA;">{{parsedUrl.host}}</div>
-				{{#if title}}
-					<div>
-						<a href="{{url}}" rel="noopener noreferrer" target="{{target}}"><strong>{{{title}}}</strong></a>
-					</div>
-				{{/if}}
-				<div>{{description}}</div>
-			</div>
-		</blockquote>
-	{{/if}}
-</template>
