--- apps/meteor/app/oembed/client/oembedVideoWidget.html
@@ -1,15 +1,0 @@
-<template name="oembedVideoWidget">
-{{#if parsedUrl}}
-	<blockquote class="message-oembed background-transparent-darker-before">
-		<div><a href="{{url}}" rel="noopener noreferrer" target="_blank">{{parsedUrl.host}}</a></div>
-		<span>{{title}}</span>
-		{{> collapseArrow collapsedMedia=collapsedMediaVar}} <br>
-		{{#unless collapsed}}
-			<video controls class="inline-video">
-				<source src="{{url}}" rel="noopener noreferrer" type="{{contentType}}">
-				Your browser does not support the video element.
-			</video>
-		{{/unless}}
-	</blockquote>
-{{/if}}
-</template>
