--- apps/meteor/app/oembed/client/oembedYoutubeWidget.html
@@ -1,12 +1,0 @@
-<template name="oembedYoutubeWidget">
-	{{#if parsedUrl}}
-		<blockquote class="message-oembed background-transparent-darker-before">
-			<a href="{{url}}" rel="noopener noreferrer" target="_blank">{{parsedUrl.host}}</a>
-			{{> collapseArrow collapsedMedia=collapsedMediaVar}} <br>
-			{{#unless collapsed}}
-				<iframe width="270" height="200" src="{{meta.twitterPlayer}}" rel="noopener noreferrer" frameborder="0" allowfullscreen></iframe><br>
-				{{{meta.description}}}
-			{{/unless}}
-		</blockquote>
-	{{/if}}
-</template>
