+++ apps/meteor/app/ui-utils/client/index.ts
@@ -15,4 +15,3 @@ export { messageProperties } from '../lib/MessageProperties';
 export { MessageTypes, MessageType } from '../lib/MessageTypes';
 export { Message } from '../lib/Message';
 export { openRoom } from './lib/openRoom';
-export * from './lib/collapseArrow';
