+++ apps/meteor/client/importPackages.ts
@@ -31,7 +31,6 @@ import '../app/message-attachments/client';
 import '../app/message-mark-as-unread/client';
 import '../app/nextcloud/client';
 import '../app/oauth2-server-config/client';
-import '../app/oembed/client';
 import '../app/otr/client';
 import '../app/push-notifications/client';
 import '../app/apps/client';
