+++ apps/meteor/client/views/room/MessageList/hooks/useParentMessage.ts
@@ -2,7 +2,7 @@ import type { IMessage } from '@rocket.chat/core-typings';
 import type { UseQueryResult } from '@tanstack/react-query';
 import { useQuery } from '@tanstack/react-query';
 
-import { findParentMessage } from '../../../../../app/ui-message/client/messageThread';
+import { findParentMessage } from '../../../../../app/ui-message/client/findParentMessage';
 
 export const useParentMessage = (mid: IMessage['_id']): UseQueryResult<IMessage> =>
 	useQuery(['parent-message', { mid }], async () => findParentMessage(mid));
