+++ apps/meteor/definition/externals/meteor/templating.d.ts
@@ -1,7 +1,7 @@
 import 'meteor/templating';
 import type { Blaze } from 'meteor/blaze';
 import type { ReactiveVar } from 'meteor/reactive-var';
-import type { IMessage, IRoom, ISubscription, IUser } from '@rocket.chat/core-typings';
+import type { IMessage, IRoom } from '@rocket.chat/core-typings';
 
 declare module 'meteor/blaze' {
 	namespace Blaze {
@@ -44,13 +44,6 @@ declare module 'meteor/templating' {
 		_noMatch: BlazeTemplate;
 		authorize: BlazeTemplate;
 		oauth404: BlazeTemplate;
-		oembedBaseWidget: BlazeTemplate;
-		oembedAudioWidget: BlazeTemplate;
-		oembedFrameWidget: BlazeTemplate;
-		oembedImageWidget: BlazeTemplate;
-		oembedUrlWidget: BlazeTemplate;
-		oembedVideoWidget: BlazeTemplate;
-		oembedYoutubeWidget: BlazeTemplate;
 		icon: BlazeTemplate<{
 			block?: string;
 			icon: string;
@@ -93,32 +86,6 @@ declare module 'meteor/templating' {
 		>;
 		error: BlazeTemplate;
 		loading: BlazeTemplate;
-		message: BlazeTemplate<{
-			msg: Omit<IMessage, 'groupable'> & { searchedText?: string; customClass?: string; actionContext?: string; groupable?: boolean };
-			room?: Omit<IRoom, '_updatedAt' | 'lastMessage'>;
-			subscription?: Pick<ISubscription, 'name' | 'autoTranslate' | 'rid' | 'tunread' | 'tunreadUser' | 'tunreadGroup'>;
-			settings: {
-				translateLanguage?: unknown;
-				autoImageLoad?: unknown;
-				saveMobileBandwidth?: unknown;
-				collapseMediaByDefault?: unknown;
-				showreply?: unknown;
-				showReplyButton?: unknown;
-				hasPermissionDeleteMessage?: unknown;
-				hasPermissionDeleteOwnMessage?: unknown;
-				hideRoles?: unknown;
-				UI_Use_Real_Name?: unknown;
-				Chatops_Username?: unknown;
-				AutoTranslate_Enabled?: unknown;
-				Message_AllowEditing?: unknown;
-				Message_AllowEditing_BlockEditInMinutes?: unknown;
-				API_Embed?: unknown;
-				Message_GroupingPeriod?: unknown;
-			};
-			u?: Partial<IUser>;
-			[key: string]: any;
-		}>;
-		messageThread: BlazeTemplate;
 		messagePopup: BlazeTemplate;
 		messagePopupChannel: BlazeTemplate;
 		messagePopupConfig: BlazeTemplate<{
@@ -134,7 +101,6 @@ declare module 'meteor/templating' {
 			getInput: () => HTMLTextAreaElement | null;
 		}>;
 		messagePopupUser: BlazeTemplate;
-		collapseArrow: BlazeTemplate;
 		rc_modal: BlazeTemplate;
 		popout: BlazeTemplate;
 		popover: BlazeTemplate;
